Improve LDAP login handling
Turns out that using ldap3.Connection() in a with…as block implicitly binds it, which is not what I want. The implicit bind throws a generic exception when the password is wrong, but the explicit bind returns False when the password/username are wrong (and throws when other things are wrong).